fix minor error
