+++ common/helper.go
@@ -4,7 +4,6 @@ import (
 	"encoding/binary"
 	"fmt"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/quad/quad-node/account"
 	"github.com/quad/quad-node/crypto/blake2b"
 	"log"
 	"time"
@@ -128,14 +127,6 @@ func ExtractKeysOfList(m map[[2]byte][][]byte) [][2]byte {
 	return keys
 }
 
-func ExtractKeysFromMapAccounts(m map[[common.AddressLength]byte]account.Account) [][common.AddressLength]byte {
-	keys := [][common.AddressLength]byte{}
-	for k, _ := range m {
-		keys = append(keys, k)
-	}
-	return keys
-}
-
 func IsInKeys2Byte(m map[[2]byte][]byte, searchKey string) bool {
 	keys := ExtractKeys(m)
 	return ContainsKey(keys, searchKey)
@@ -144,11 +135,6 @@ func IsInKeys2Byte(m map[[2]byte][]byte, searchKey string) bool {
 func IsInKeysOfList(m map[[2]byte][][]byte, searchKey [2]byte) bool {
 	keys := ExtractKeysOfList(m)
 	return ContainsKeyOfList(keys, searchKey)
-}
-
-func IsInKeysOfMapAccounts(m map[[common.AddressLength]byte]account.Account, searchKey [common.AddressLength]byte) bool {
-	keys := ExtractKeysFromMapAccounts(m)
-	return ContainsKeyInMap(keys, searchKey)
 }
 
 func BytesToLenAndBytes(b []byte) []byte {
